+++ Service/AddOn/ReadmeBuilder.php
@@ -120,6 +120,18 @@ class ReadmeBuilder extends AbstractService
             $this->applyDefaultOrder($finder);
             $this->applyIdentifierSpecificConditions($finder);
 
+            $sortBy = null;
+            if (is_array($dataKey))
+            {
+                $sortBy = $dataKey['sortBy'] ?? null;
+                $dataKey = $dataKey['keyedBy'];
+
+                if ($sortBy !== null)
+                {
+                    $this->applyCustomOrder($finder, $sortBy);
+                }
+            }
+
             $data[$dataKey] = $finder->fetch();
         }
 
@@ -201,6 +213,8 @@ class ReadmeBuilder extends AbstractService
     }
 
     /**
+     * @version 1.5.0
+     *
      * @return array<string, string>
      */
     protected function getFinderAndDataMap() : array
@@ -210,7 +224,13 @@ class ReadmeBuilder extends AbstractService
             'XF:BbCodeMediaSite' => 'bb_code_media_sites',
             'XF:CronEntry' => 'cron_entries',
             'XF:Option' => 'options',
-            'XF:Permission' => 'permissions',
+            'XF:Permission' => [
+                'keyedBy' => 'permissions',
+                'sortBy' => [
+                    'Interface.display_order' => 'ASC',
+                    'display_order' => 'ASC'
+                ]
+            ],
             'XF:StyleProperty' => 'style_properties',
             'XF:WidgetDefinition' => 'widget_definitions',
             'XF:WidgetPosition' => 'widget_positions',
@@ -352,6 +372,34 @@ class ReadmeBuilder extends AbstractService
         }
 
         return $finder;
+    }
+
+    /**
+     * @since 1.5.0
+     *
+     * @param Finder $finder
+     * @param mixed $sortBy
+     *
+     * @return void
+     */
+    protected function applyCustomOrder(Finder $finder, $sortBy)
+    {
+        if (is_null($sortBy))
+        {
+            return;
+        }
+
+        if (is_array($sortBy))
+        {
+            foreach ($sortBy AS $column => $direction)
+            {
+                $finder->order($column, $direction);
+            }
+        }
+        else
+        {
+            $finder->order($sortBy);
+        }
     }
 
     /**