+++ CHANGELOG.md
@@ -3,7 +3,8 @@ CHANGELOG
 
 ## 1.5.0 (`1050070`)
 
-- **New:** Add tck-devtools:finder-class-properties command to apply class properties to type hint relations in finder (#126)
+- **New:** Add `tck-devtools:finder-class-properties` command to apply class properties to type hint relations in finder (#126)
+- **Fix:** Permission interface groups are not sorted correctly when building readme (#125)
 
 ## 1.4.3 (`1040370`)